+++ SConstruct
@@ -2,4 +2,4 @@
 SConscript('board/SConscript')
 
 # test files
-SConscript('tests/safety/SConscript')
+SConscript('tests/libpanda/SConscript')
