+++ tests/libpanda/SConscript
@@ -1,0 +1,11 @@
+env = Environment(
+  CC='gcc',
+  CFLAGS=[
+    '-nostdlib',
+    '-fno-builtin',
+    '-std=gnu11',
+  ],
+  CPPPATH=[".", "#board/"],
+)
+
+env.SharedLibrary("libpanda.so", ["safety_helpers.c"])
