+++ tests/libpanda/safety_helpers.c
@@ -164,10 +164,8 @@ void init_tests(void){
   ts_steer_req_mismatch_last = 0;
   valid_steer_req_count = 0;
   invalid_steer_req_count = 0;
-}
 
-void init_tests_honda(void){
-  init_tests();
+  // car-specific stuff
   honda_fwd_brake = false;
 }
