+++ tests/safety/test.sh
@@ -4,8 +4,6 @@
 # Make sure test fails if one HW_TYPE fails
 set -e
 
-scons -u
-
 for hw_type in {0..7}; do
   echo "Testing HW_TYPE: $hw_type"
   HW_TYPE=$hw_type python -m unittest discover .
