+++ tests/safety/test_volkswagen_mqb.py
@@ -2,7 +2,7 @@
 import unittest
 import numpy as np
 from panda import Panda
-from panda.tests.safety import libpandasafety_py
+from panda.tests.libpanda import libpanda_py
 import panda.tests.safety.common as common
 from panda.tests.safety.common import CANPackerPanda
 
@@ -143,7 +143,7 @@ class TestVolkswagenMqbStockSafety(TestVolkswagenMqbSafety):
 
   def setUp(self):
     self.packer = CANPackerPanda("vw_mqb_2010")
-    self.safety = libpandasafety_py.libpandasafety
+    self.safety = libpanda_py.libpanda
     self.safety.set_safety_hooks(Panda.SAFETY_VOLKSWAGEN_MQB, 0)
     self.safety.init_tests()
 
@@ -165,7 +165,7 @@ class TestVolkswagenMqbLongSafety(TestVolkswagenMqbSafety):
 
   def setUp(self):
     self.packer = CANPackerPanda("vw_mqb_2010")
-    self.safety = libpandasafety_py.libpandasafety
+    self.safety = libpanda_py.libpanda
     self.safety.set_safety_hooks(Panda.SAFETY_VOLKSWAGEN_MQB, Panda.FLAG_VOLKSWAGEN_LONG_CONTROL)
     self.safety.init_tests()
