+++ tests/safety_replay/replay_drive.py
@@ -2,12 +2,12 @@
 import argparse
 import os
 
-from panda.tests.safety import libpandasafety_py
+from panda.tests.libpanda import libpanda_py
 from panda.tests.safety_replay.helpers import package_can_msg, init_segment
 
 # replay a drive to check for safety violations
 def replay_drive(lr, safety_mode, param, alternative_experience, segment=False):
-  safety = libpandasafety_py.libpandasafety
+  safety = libpanda_py.libpanda
 
   err = safety.set_safety_hooks(safety_mode, param)
   assert err == 0, "invalid safety mode: %d" % safety_mode
